+++ backend/src/baserow/config/settings/base.py
@@ -1,9 +1,11 @@
 import datetime
 import importlib
 import os
+import re
 from decimal import Decimal
 from pathlib import Path
 from urllib.parse import urljoin, urlparse
+from ipaddress import ip_network
 
 import dj_database_url
 from celery.schedules import crontab
@@ -564,6 +566,27 @@ WEBHOOKS_MAX_RETRIES_PER_CALL = 8
 WEBHOOKS_MAX_PER_TABLE = 20
 WEBHOOKS_MAX_CALL_LOG_ENTRIES = 10
 WEBHOOKS_REQUEST_TIMEOUT_SECONDS = 5
+BASEROW_WEBHOOKS_ALLOW_PRIVATE_ADDRESS = bool(
+    os.getenv("BASEROW_WEBHOOKS_ALLOW_PRIVATE_ADDRESS", False)
+)
+BASEROW_WEBHOOKS_IP_BLACKLIST = [
+    ip_network(ip.strip())
+    for ip in os.getenv("BASEROW_WEBHOOKS_IP_BLACKLIST", "").split(",")
+    if ip.strip() != ""
+]
+BASEROW_WEBHOOKS_IP_WHITELIST = [
+    ip_network(ip.strip())
+    for ip in os.getenv("BASEROW_WEBHOOKS_IP_WHITELIST", "").split(",")
+    if ip.strip() != ""
+]
+BASEROW_WEBHOOKS_URL_REGEX_BLACKLIST = [
+    re.compile(url_regex.strip())
+    for url_regex in os.getenv("BASEROW_WEBHOOKS_URL_REGEX_BLACKLIST", "").split(",")
+    if url_regex.strip() != ""
+]
+BASEROW_WEBHOOKS_URL_CHECK_TIMEOUT_SECS = int(
+    os.getenv("BASEROW_WEBHOOKS_URL_CHECK_TIMEOUT_SECS", "10")
+)
 
 # ======== WARNING ========
 # Please read and understand everything at: